+++ lending/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v0.3.0] lending
 ### Releases
 - [NPM v0.3.0] https://www.npmjs.com/package/lending/v/0.3.0 - lending
+
+## 2023-11-20 01:58:00
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.1] lending
+### Releases
+- [NPM v0.3.1] https://www.npmjs.com/package/lending/v/0.3.1 - lending